% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-logit-reg.R
\name{check_inputs}
\alias{check_inputs}
\title{Assertion of inputs for logistic regression}
\usage{
check_inputs(design, response, coefs)
}
\arguments{
\item{design}{design matrix of independent variables}

\item{response}{numeric vector of a dependent variable}

\item{coefs}{numeric vector for initial values of coefficients with which
the log-likelihood must be optimized}
}
\value{
vector of coefficients if assertin successful; otherwise error
}
\description{
Assert that input data for logistic regression model follow necessary
conditions for estimation
}
