% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg-helper-methods.R
\name{coef.logitreg}
\alias{coef.logitreg}
\title{method \code{coef} for class \code{logitreg}}
\usage{
\method{coef}{logitreg}(object, ...)
}
\arguments{
\item{object}{object of class \code{logitreg}.}

\item{...}{further arguments.}
}
\value{
a vector of estimated coefficients
}
\description{
Returns estimated coefficients from the \link{logitreg} object.
}
