% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{fit_logitreg}
\alias{fit_logitreg}
\alias{fit_logitreg.formula}
\alias{fit_logitreg.default}
\title{Basic fitting function for a logit model.}
\usage{
fit_logitreg(design = NULL, ...)

\method{fit_logitreg}{formula}(
  design = NULL,
  data = NULL,
  hessian = TRUE,
  check_separable = TRUE,
  ...
)

\method{fit_logitreg}{default}(
  design = NULL,
  response = NULL,
  hessian = TRUE,
  check_separable = TRUE,
  ...
)
}
\arguments{
\item{design}{design matrix, numeric matrix or model formula, formula}

\item{\dots}{additional arguments passed to the methods or \code{optim}'s control
argument}

\item{data}{data.frame}

\item{hessian}{should the hessian be computed? logical}

\item{check_separable}{should a test of linearly separability pe performed?
logical}

\item{response}{response vector, integer vector}
}
\value{
an object of class \code{logitreg}, i.e., a list including the estimated
coefficients, the fitted values, the data used for fitting, the vcov (if
hessian = TRUE, otherwise a matrix of NA's), the log likelihood of the
fitted model, the number of parameters estimated, the convergence code of
optim, and the number of gradient evaluations used during optimization.
Several basic methods are available, namely \code{coef}, \code{vcov}, \code{logLik},
\code{summary}, \code{plot}, \code{predict}, and \code{fitted}.
}
\description{
Estimate the parameter vector under a logit model providing a design matrix
and a response vector. Linear separability of the data is checked solving a
linear program via Rglpk (if \code{check_separable = TRUE}, default).
}
\examples{
data <- sim_data(seed = 123, dataframe = TRUE)
glmx <- fit_logitreg(response ~ X1 + X2 + X3, data = data)
summary(glmx)
}
