% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg-helper-methods.R
\name{fitted.logitreg}
\alias{fitted.logitreg}
\title{method \code{fitted} for class \code{logitreg}}
\usage{
\method{fitted}{logitreg}(object, ...)
}
\arguments{
\item{object}{object of class \code{logitreg}.}

\item{...}{further arguments.}
}
\value{
a vector of fitted values
}
\description{
Returns fitted values from the \link{logitreg} object.
}
