% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{is_separable}
\alias{is_separable}
\title{Test for linear separability.}
\usage{
is_separable(design, response)
}
\arguments{
\item{design}{design matrix, numeric matrix or model formula, formula}

\item{response}{response vector, integer vector}
}
\value{
a logical whether the data is linear separable
}
\description{
Test whether data is linear separable by solving a linear program via Rglpk.
}
