% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{logistic}
\alias{logistic}
\title{Compute the logistic transformation}
\usage{
logistic(x)
}
\arguments{
\item{x}{value to transform, numeric}
}
\value{
logistic transformed value, numeric
}
\description{
Simple wrapper around \link[stats:plogis]{stats::plogis}.
}
