% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{logitreg}
\alias{logitreg}
\alias{logitreg.formula}
\alias{logitreg.default}
\alias{fit_logitreg}
\title{Fit logistic regression by minimising negative log likelihood.}
\usage{
logitreg(x, y, init_coefs, max_iter, ...)

\method{logitreg}{formula}(x, y, ...)

\method{logitreg}{default}(x, y, init_coefs = rnorm(ncol(x)),
  max_iter = 1000, ...)

fit_logitreg(design, response, init_coefs, max_iter, ...)
}
\arguments{
\item{x}{formula or design matrix}

\item{y}{design matrix or response}

\item{init_coefs}{optional numeric vector of initial coefficient values}

\item{max_iter}{maximal number of allowed iterations}

\item{...}{optional \code{\link[=optim]{optim()}} arguments}

\item{design}{design matrix with n rows and q columns}

\item{response}{response vector of length n}
}
\value{
logitreg object, see \code{\link[=new_logitreg]{new_logitreg()}}
}
\description{
Fit logistic regression by minimising negative log likelihood.
}
\examples{
formula <- y ~ x
data <- data.frame(x = c(0, 1), y = c(0, 1))
logitreg(formula, data)
design <- matrix(c(1, 1, 0, 1), nrow = 2)
response <- c(0, 1)
logitreg(design, response)
}
