% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{neg_loglik}
\alias{neg_loglik}
\alias{neg_loglik_deriv}
\title{Compute the negative log likelihood of some data under a logit model or its
gradient.}
\usage{
neg_loglik(coefs, design, response)

neg_loglik_deriv(coefs, design, response)
}
\arguments{
\item{coefs}{parameter vector, numeric vector}

\item{design}{design matrix, numeric matrix}

\item{response}{response vector, numeric vector}
}
\value{
negative log likelihood, numeric vector

gradient of the negative log likelihood, numeric vector
}
\description{
Based on a parameter vector, design matrix and a response vector, the
negative log likelihood under the logit model or its gradient is computed.
}
