% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{new_logitreg}
\alias{new_logitreg}
\title{Constructor for logitreg results}
\usage{
new_logitreg(design, response, coefs)
}
\arguments{
\item{design}{design matrix with n rows and q columns}

\item{response}{response vector of length n}

\item{coefs}{optimal coefficients}
}
\value{
list of
\code{coefficients}: the regression coefficients,
\code{fitted}: the estimated probabilities
\code{data}: the original data (\code{design} & \code{response}) used for the fit.
}
\description{
Constructor for logitreg results
}
