% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{plot.logitreg}
\alias{plot.logitreg}
\title{Plot ROC curve}
\usage{
\method{plot}{logitreg}(...)
}
\arguments{
\item{...}{arguments to pass to ROCR::plot.}
}
\description{
Plot ROC curve
}
\examples{
data <- logitreg::sim_data(300, seed = 321, numerics = 2, dataframe = TRUE)
model <- logitreg(response ~ ., data = data)
pred <- ROCR::prediction(predict(model,
  type = "response"
), data$response)
perf <- ROCR::performance(pred, "acc")
plot(perf)
}
