% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{predict.logitreg}
\alias{predict.logitreg}
\title{Predict function for logitreg object}
\usage{
\method{predict}{logitreg}(object, type = "link", ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{type}{the type of prediction required.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
a vector of preditions.
}
\description{
Predict function for logitreg object
}
