% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{sim_data}
\alias{sim_data}
\title{Generate synthetic data sets for tests}
\usage{
sim_data(
  n = 1500L,
  numerics = 3L,
  factors = 0L,
  seed = NULL,
  dataframe = FALSE
)
}
\arguments{
\item{n}{number of observations, integer}

\item{numerics}{number of numeric variables, integer}

\item{factors}{number of factor variables, integer}

\item{seed}{seed for the RNG, numeric}

\item{dataframe}{should a dataframe be returned consisting of the response
and the variables (dropping the intercept column and appending the coefs
as an attribute) or should everything be returned as a list (design,
repsonse, coefs)? logical}
}
\value{
list of the design, response and the coefs, or a dataframe including
the response and the design (minus the intercept column) with the coefs
appended as an attribute
}
\description{
Generates synthetic data sets for logistic regression.
}
