% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg-helper-methods.R
\name{summary.logitreg}
\alias{summary.logitreg}
\title{method \code{summary} for class \code{logitreg}}
\usage{
\method{summary}{logitreg}(object, ...)
}
\arguments{
\item{object}{object of class \code{logitreg}.}

\item{...}{further arguments.}
}
\value{
a list of summary statistics:
\itemize{
\item \code{Call} formula of the model
\item \code{Coefficients} vector of estimated coefficients
\item \code{Confusion} confusion table for true and fitted values
}
}
\description{
Produces summary of results for the \link{logitreg} object.
}
