% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-helpers-logitreg.R
\name{check_inputs}
\alias{check_inputs}
\title{Assertion of inputs for logistic regression}
\usage{
check_inputs(response, design, coefs)
}
\arguments{
\item{response}{vector of responses}

\item{design}{design matrix for the model}

\item{coefs}{start values for coefficients}
}
\value{
vector of coefficients if assertion successful; otherwise error
}
\description{
Assert that input data for logistic regression model follow necessary
conditions for estimation
}
