% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logitreg.R
\name{fit_logitreg}
\alias{fit_logitreg}
\title{Fitting logistic regresion}
\usage{
fit_logitreg(design, response, coefs_init = NULL, ...)
}
\arguments{
\item{design}{Numeric matrix containing the (scaled) design matrix.}

\item{response}{Numeric vector containing the (scaled) response.}

\item{coefs_init}{Numeric vector containing initial coefficientsfor the optimization.
If `NULL` then the coefficients are drawn out of a uniform distribution over the intervall [-3, 3]}

\item{...}{For \code{\link[stats]{optim}} : Additonal arguments for likelihood maximization procedure.}
}
\value{
A list containing the regression coefficients, the estimated probabilites P(Y = 1) and
the (original) data
}
\description{
Fitting procedure for logistic regression given a design matrix and a response vector.
}
