% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-logit-reg.R
\name{fit_logitreg}
\alias{fit_logitreg}
\title{Fitting Logistic Regression Model}
\usage{
fit_logitreg(design, response, coefs = NULL, ...)
}
\arguments{
\item{design}{design matrix of independent variables}

\item{response}{numeric vector of a dependent variable}

\item{coefs}{numeric vector for initial values of coefficients with which
the log-likelihood must be optimized}

\item{...}{further arguments to be passed, see \link{optim}}
}
\value{
a list with components:
\itemize{
\item $coefficients$   best coefficients found to fit given data
\item $fitted$         estimated probabilities
\item $data$           a list with initial $design$ and $response$ data for which
the log-likelihood is optimized
}
}
\description{
$fit_logitreg$ is used to fit a logistic regression model minimizing
the negative log-likelihood of the given data.
}
