% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logitreg.R
\name{fit_logitreg}
\alias{fit_logitreg}
\title{Fitting logistic regression models}
\usage{
fit_logitreg(design, response, ...)
}
\arguments{
\item{design}{design matrix (n,p)}

\item{response}{response vector (n)}

\item{...}{optional arguments for the \code{optim} function for the estimation}
}
\value{
list of:
\itemize{
\item \code{coefficients} estimated MLE coefficients
\item \code{fitted} estimated probabilities
\item \code{data} list of reponse and design matrix used for the fit
}
}
\description{
fit_logitreg is used to fit logistic regression models. The estimation
method is Maximum likelihood.
}
\details{
observations with NA's will be omitted
}
\examples{
fit_logitreg(
  stats::model.matrix(~., data = mtcars[, -8]),
  mtcars$vs
)
}
