% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg-helpers.R
\name{logistic}
\alias{logistic}
\title{Logistic Distribution Function}
\usage{
logistic(x)
}
\arguments{
\item{x}{Numeric vector containing the observations.}
}
\value{
Numeric vector containing the values of the distribution function for the given observations.
}
\description{
Calculates the values of the logistic distribution function for given observations.
It is assumed that the data follow a distribution function that has location and scale values of 0.
}
