% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{logistic}
\alias{logistic}
\title{Squash score predictions between 0 and 1 using the sigmoid function.}
\usage{
logistic(x)
}
\arguments{
\item{x}{scores in real numbers}
}
\value{
probability predictions
}
\description{
Squash score predictions between 0 and 1 using the sigmoid function.
}
