% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-logit-reg.R
\name{logistic}
\alias{logistic}
\title{Computation of values of distribution function for the logistic distribution}
\usage{
logistic(x)
}
\arguments{
\item{x}{single numeric value}
}
\value{
distribution function
}
\description{
Compute distribution function of the logistic distribution for the given $x$
}
\seealso{
\link[stats:Logistic]{stats::Logistic}
}
