% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{logistic}
\alias{logistic}
\title{The logistic distribution function}
\usage{
logistic(x)
}
\arguments{
\item{x}{vector of quantiles.}
}
\value{
cummulativ distribution function of x.
}
\description{
The logistic distribution function
}
