% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_logitreg.R
\name{logitreg}
\alias{logitreg}
\alias{logitreg.default}
\alias{logitreg.formula}
\title{Generic function for logistic regression models}
\usage{
logitreg(design = NULL, response = NULL, coefs_init = NULL, ...)

\method{logitreg}{default}(design, response, coefs_init = NULL, ...)

\method{logitreg}{formula}(design, response, coefs_init = NULL, ...)
}
\arguments{
\item{design}{Numeric matrix containing the design matrix.}

\item{response}{Numeric vector containing the response.}

\item{coefs_init}{Numeric vector containing initial coefficientsfor the optimization.
If `NULL` then the coefficients are drawn out of a uniform distribution over the intervall [-3, 3]}

\item{...}{For \code{\link[stats]{optim}} : Additonal arguments for likelihood maximization procedure.}
}
\value{
A list of class `logitreg` containing the regression coefficients, the estimated probabilites P(Y = 1) and
the (original) data
}
\description{
`logitreg` is a generic function for logistic regression models. It estimates the coefficients
given a design matrix and a response vector. The function invokes particular methods which depend
on the \code{\link[base]{class}} of the first argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for logitreg

\item \code{formula}: Takes model formula and data.frame as arguments
}}

