% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-logitreg.R
\name{logitreg}
\alias{logitreg}
\alias{logitreg.formula}
\alias{logitreg.default}
\title{\code{logitreg} generic & methods}
\usage{
logitreg(object, data, coefs = NULL, ...)

\method{logitreg}{formula}(object, data, coefs = NULL, ...)

\method{logitreg}{default}(object, data, coefs = NULL, ...)
}
\arguments{
\item{object}{an object of class "\link{formula}", see Details.}

\item{data}{a data frame containing the variables in the model as described
in the formula, see Details.}

\item{coefs}{numeric vector for initial values of coefficients with which
the log-likelihood must be optimized.}

\item{...}{further arguments to be used for optimization; see \link{optim}.}
}
\value{
an object of class \code{logitreg} with the following components:
\itemize{
\item \code{coefficients} best coefficients found to fit given data
\item \code{fitted} predictions with the best coefficients for the given data
\item \code{data} a list with initial \code{design} matrix and
\code{response} for which the log-likelihood is optimized
\item \code{formula} initial \code{formula} used to fit the model
}
}
\description{
The function is used to fit a logistic regression model minimizing
the negative log-likelihood of the given data.
}
\details{
Alternatively, one can supply a vector of responses as an
\code{object} and a design matrix (see \link{model.matrix}) as \code{data}. In this case
the function will return a list of estimated coefficients, fitted values
and initial data used for estimation. This is not advised as the returned
object will be a simple list and not an object of class \code{logitreg}.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: used if a \code{formula} and a \code{data.frame} are supplied

\item \code{default}: used if a response vector and a design matrix are supplied
}}

