% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{logitreg}
\alias{logitreg}
\alias{logitreg.default}
\alias{logitreg.formula}
\title{Fit logistic regression}
\usage{
logitreg(design, ...)

\method{logitreg}{default}(design, response, ...)

\method{logitreg}{formula}(design, data, ...)
}
\arguments{
\item{design}{a design matrix or a formula}

\item{...}{method specific arguments}
}
\value{
object of class logitreg
}
\description{
For fitting \code{fit_logitreg()} is used.
}
\seealso{
\link{ fit_logitreg() }
}
