% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{logitreg}
\alias{logitreg}
\alias{logitreg.default}
\alias{logitreg.formula}
\title{logitreg: logistic regression function.}
\usage{
logitreg(object, ...)

\method{logitreg}{default}(object, response, ...)

\method{logitreg}{formula}(formula, data, ...)
}
\arguments{
\item{object}{design matrix.}

\item{...}{other arguments used by \code{optim}.}

\item{response}{binary dependent variable.}

\item{formula}{an object of class "formula".
(or one that can be coerced to that class).}

\item{data}{a data frame.}
}
\value{
A list of class "logitreg".
containing the optimized coefficients,
fitted and predicted response probabilities
and the original design matrix.
}
\description{
Finds the MLE of regression coefficients using numerical optimization.
}
\details{
Fit logistic regression
}
