% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg-helpers.R
\name{neg_loglik}
\alias{neg_loglik}
\alias{neg_loglik_deriv}
\title{Negative Log-Likelihood for logistic regression}
\usage{
neg_loglik(coefs, design, response)

neg_loglik_deriv(coefs, design, response)
}
\arguments{
\item{coefs}{Numeric vector containing the coefficients for the logistic regression.}

\item{design}{Numeric matrix containing the design matrix.}

\item{response}{Numeric vector containing the response.}
}
\value{
Numeric vector containing the values of the distribution function for the given observations.
}
\description{
Calculates the value of the logistic distribution or its gradient given the coefficients, design
matrix and the response vector.
}
\details{
Given the coefficients, the design matrix and the response vector, neg_loglik and neg_loglik_deriv
calculate the negative log-likelihood and its gradient, respectively.
}
