% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{neg_loglik}
\alias{neg_loglik}
\alias{neg_loglik_deriv}
\title{Compute the negative log likelihood of logistic regression.}
\usage{
neg_loglik(coefs, design, response)

neg_loglik_deriv(coefs, design, response)
}
\arguments{
\item{coefs}{vector of regression coefficients}

\item{design}{design matrix with n rows and q columns}

\item{response}{response vector of length n}
}
\value{
negative log likelihood
}
\description{
Compute the negative log likelihood of logistic regression.

Compute the derivative of negative log likelihood of logistic regression.
}
\examples{
coefs <- matrix(c(-23, 47), nrow = 2)
design <- matrix(c(1, 1, 0, 1), nrow = 2)
response <- c(0, 1)
neg_loglik(coefs, design, response)
}
