% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-logit-reg.R
\name{neg_loglik}
\alias{neg_loglik}
\title{Computation of negative log-likelihood for logistic regression}
\usage{
neg_loglik(coefs, design, response)
}
\arguments{
\item{coefs}{numeric vector for initial values of coefficients with which
the log-likelihood must be optimized}

\item{design}{design matrix of independent variables}

\item{response}{numeric vector of a dependent variable}
}
\value{
negative log-likelihood for given data and coefficients
}
\description{
Compute negative log-likelihood of data based on the logistic distribution
}
