% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logitreg.R
\name{neg_loglik}
\alias{neg_loglik}
\alias{neg_loglik_deriv}
\title{Compute neg. loglikelihood and derivative (log. regression)}
\usage{
neg_loglik(coefs, design, response)

neg_loglik_deriv(coefs, design, response)
}
\arguments{
\item{coefs}{model coefficients (p)}

\item{design}{designmatrix (n,p)}

\item{response}{response vector (n)}
}
\value{
value of neg. loglikelihood/ derivative
}
\description{
functions to cumpute the negative loglikelihood and the derivative
of a logistic regression
}
