% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{neg_loglik}
\alias{neg_loglik}
\alias{neg_loglik_deriv}
\title{Negative log likelihood and negative gradient computation}
\usage{
neg_loglik(coefs, design, response)

neg_loglik_deriv(coefs, design, response)
}
\arguments{
\item{coefs}{regression parameter vector.}

\item{design}{design matrix.}

\item{response}{binary dependent variable.}
}
\value{
\code{neg_loglik} gives the negative log likelihood.

\code{neg_loglik_deriv} gives the gradient of the negative log likelihood.
}
\description{
computes the negative log likelihood
and the gradient of the logistic density
for some data points in the design matrix.
}
