% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-helpers-logitreg.R
\name{neg_loglik_deriv}
\alias{neg_loglik_deriv}
\title{Computation of the negative gradient of the log-likelihood}
\usage{
neg_loglik_deriv(response, design, coefs)
}
\arguments{
\item{response}{vector of responses}

\item{design}{design matrix for the model}

\item{coefs}{start values for coefficients}
}
\value{
negative gradient values for given coefficients
}
\description{
Compute the first negative derivative / gradient of the log-likelihood
for given data and coefficients based on the logistic distribution
}
