% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-logitreg.R
\name{new_logitreg}
\alias{new_logitreg}
\title{Constructor for class \code{logitreg}}
\usage{
new_logitreg(fit)
}
\arguments{
\item{fit}{model for which class \code{logitreg} must be set}
}
\value{
an object of class \link{logitreg}
}
\description{
Create an object of class \code{logitreg}
}
