% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_logitreg.R
\name{plot.logitreg}
\alias{plot.logitreg}
\title{Method `plot` for class `logitreg`}
\usage{
\method{plot}{logitreg}(x, ...)
}
\arguments{
\item{x}{object of class `logitreg`.}

\item{...}{additional arguments for plot function}
}
\description{
Obtains ROC-curve plot for objects of class `logitreg`
}
