% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_logitreg.R
\name{predict.logitreg}
\alias{predict.logitreg}
\title{Method `predict` for class `logitreg`}
\usage{
\method{predict}{logitreg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{object of class `logitreg`.}

\item{newdata}{optionally,  a data frame in which to look for variables with which to predict.
If omitted, the fitted predictors are used.#'}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a vector of predictions.
}
\description{
Obtains predictions for new data from fitted logistic regression model object.
}
