% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg-helper-methods.R
\name{predict.logitreg}
\alias{predict.logitreg}
\title{method \code{predict} for class \code{logitreg}}
\usage{
\method{predict}{logitreg}(object, newdata, ...)
}
\arguments{
\item{object}{object of class \code{logitreg}.}

\item{newdata}{optional data frame with new data for which predictions
should be calculated. If omitted, fitted values are returned.}

\item{...}{further arguments.}
}
\value{
a vector of predictions.
}
\description{
Predicts values for new data based on the \link{logitreg} object.
}
