% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg.R
\name{predict.logitreg}
\alias{predict.logitreg}
\title{Predict probabilities for logitreg models}
\usage{
\method{predict}{logitreg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{logitreg object}

\item{newdata}{data.frame or design matrix with new data
(default returns fitted probabilities for training data)}
}
\value{
numeric vector with probabilities of size nrow(newdata)
}
\description{
Get the predicted probabilities for a logistic regression model.
}
