% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitreg-helpers.R
\name{sim_data}
\alias{sim_data}
\title{Simulation of synthetic data sets for tests
Generates synthetic data sets for tests with q_numeric gaussian variables (~N(0,1))
and q_factor dummy variables}
\usage{
sim_data(n = 1500, numerics = 3, factors = 0, seed = NULL,
  dataframe = FALSE)
}
\arguments{
\item{n}{Number of simulated oberservations.}

\item{numerics}{Number of numeric columns in the simulated design matrix.}

\item{factors}{Number of factor columns in the simulated design matrix.}

\item{seed}{A single starting value for random number generators.}

\item{dataframe}{If `TRUE`, the default, the results are returned as a list. If `FALSE`, the
results are returned as a data.frame.}
}
\value{
A list(x, y, b): design, response, true coefs or data.frame with attribute('b')
}
\description{
Simulation of synthetic data sets for tests
Generates synthetic data sets for tests with q_numeric gaussian variables (~N(0,1))
and q_factor dummy variables
}
