% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sim_data}
\alias{sim_data}
\title{Generate synthetic data sets for tests.}
\usage{
sim_data(n = 1500, numerics = 3, factors = 0, seed = NULL,
  dataframe = FALSE)
}
\arguments{
\item{n}{number of rows}

\item{numerics}{number of numeric columns}

\item{factors}{number of factor columns}

\item{seed}{random seed}

\item{dataframe}{flag for returning dataframe or list as output}
}
\value{
list(x = design, y = response, b = true coefs)
or dataframe with attribute('b')
}
\description{
Generate synthetic data sets for tests.
}
