% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-synthetic-data.R
\name{sim_data}
\alias{sim_data}
\title{Simulation of synthetic data}
\usage{
sim_data(n = 1500, numerics = 3, factors = 0, seed = NULL,
  dataframe = FALSE)
}
\arguments{
\item{n}{number of rows}

\item{numerics}{number of numeric covariates}

\item{factors}{number of factorisl covariates}

\item{seed}{seed for reproducibility}

\item{dataframe}{logical; if \code{TRUE} dataframe as output, otherwise matrix}
}
\value{
list with three components:
\itemize{
\item \code{design} model matrix for covariates
\item \code{response} numeric vector of responses
\item \code{coefs} true coefficients with which synthetic data are simulated
}
}
\description{
Simulate synthetic data for tests
}
