% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{sim_data}
\alias{sim_data}
\title{Generate synthetic data sets for tests.}
\usage{
sim_data(n = 1500, numerics = 3, factors = 0, seed = NULL, dataframe = FALSE)
}
\arguments{
\item{n}{number of observations to create.}

\item{numerics}{number of gaussian variables (~N(0,1)).}

\item{factors}{number of dummy variables.}

\item{seed}{RNG seed.}

\item{dataframe}{logical value
specifying if the returned object is a list or a datafram.}
}
\value{
list(x, y, b): design, response, true
coefs or data.frame with attribute('b').
}
\description{
Generate synthetic data sets for tests.
}
